addded camera movement

+++ programm/color/color.h
@@ -9,14 +9,14 @@ struct ColorRGB
 {
  public:
 	ColorRGB() = default;
-	ColorRGB(int R_,int G_,int B_);
+	ColorRGB(float R_, float G_, float B_);
 	ColorRGB operator*(float value);
 	ColorRGB operator+(const ColorRGB& color);
 	ColorRGB operator*(const ColorRGB& color);
 	void normalize();
-	int R;
-	int G;
-	int B;
+	float R;
+	float G;
+	float B;
 };
 
 #endif //LAB_03_CG_COURSE_PROGRAMM_COLOR_COLOR_H_
